// src/app/model/vehicle/vehicle.ts
export interface Vehicle {
  imei: string;
  alias: string;
  automotor_plate: string;
  group_name: string;
  longitude: number;
  latitude: number;
  speed: number;
  device_status: string;
  ignition_state: number;
  movement_state: number;
  generated_time: string;
  message_time: string;
  model_name: string;
  device_type_name: string;
  event_id: number;
  battery_level: number;
  odometer: number;
  iconUrl: string;
}